Update and rename src/components/Admin/Form/FormMoneda/FormMoneda.js to src/componentes/Administración/Formulario/FormMoneda/FormMoneda.js
